+++ spring05-library-jdbc/src/main/java/ru/otus/spring/library/springjdbc/dao/GenreDao.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface GenreDao {
 
-    Genre getById(int id);
+    Genre getById(long id);
 
     Genre getByName(String genreName);
